requirements.txt
pydantic[email,ai]==2.9.2
python-dotenv==1.0.1
openai==1.50.2
